requirements.txt
# Backend Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# LLM & RAG (as per documentation)
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.20

# Embeddings & Vector Store
chromadb==0.4.22
sentence-transformers==2.3.1

# PDF Processing
pymupdf==1.23.8
pypdf==3.17.4

# Utilities
python-dotenv==1.0.0
numpy==1.26.3